renamed modules, updated the name change everywhere else
